+++ services/dashboard-ui/client/lite/lib/diffs/helm.test.ts
@@ -1,10 +1,14 @@
 import { describe, expect, test } from 'bun:test'
-import type { THelmPlan } from '@/types'
 import {
   mixedHelmPlan,
+  redisClusterRollbackPlan,
   vmagentSingleRemovalPlan,
-} from '../fixtures/plan-diffs/helm'
+} from '@/lib/fixtures/plan-diffs/helm'
+import type { THelmPlan } from '@/types'
 import { helmPlanDiff } from './helm'
+
+const occurrences = (text: string, needle: string) =>
+  text.split(needle).length - 1
 
 describe('helmPlanDiff', () => {
   test('normalizes direct content and summary counts', () => {
@@ -15,26 +19,60 @@ describe('helmPlanDiff', () => {
       'create',
       'delete',
     ])
+    expect(result.sections.map(({ title }) => title)).toEqual([
+      'my-app',
+      'my-app-svc',
+      'my-app-cache',
+    ])
     expect(result.summary).toMatchObject({
       create: 1,
       update: 1,
       delete: 1,
     })
-    expect(result.sections[0]?.before).toContain(
-      'image: example.com/payments:1.2.0'
-    )
-    expect(result.sections[0]?.after).toContain(
-      'image: example.com/payments:1.3.0'
-    )
+    expect(result.sections[0]?.before).toContain('image: my-app:1.2.0')
+    expect(result.sections[0]?.after).toContain('image: my-app:1.3.0')
+    expect(result.sections[1]?.before).toBe('')
+    expect(result.sections[1]?.after).toContain('kind: Service')
+    expect(result.sections[2]?.before).toContain('image: redis:6.2')
+    expect(result.sections[2]?.after).toBe('')
   })
 
   test('builds before and after content from entry arrays', () => {
     const [section] = helmPlanDiff(vmagentSingleRemovalPlan).sections
 
-    expect(section?.before).toContain('--legacy-endpoint=:8429')
-    expect(section?.after).not.toContain('--legacy-endpoint=:8429')
-    expect(section?.before).toContain('--logger-format=json')
-    expect(section?.after).toContain('--logger-format=json')
+    expect(section?.operation).toBe('update')
+    expect(section?.before).toContain('image: victoriametrics/vmagent:v1.132.0')
+    expect(section?.after).toContain('image: victoriametrics/vmagent:v1.132.0')
+    expect(
+      occurrences(section?.before ?? '', '--remoteWrite.bearerToken=')
+    ).toBe(2)
+    expect(
+      occurrences(section?.after ?? '', '--remoteWrite.bearerToken=')
+    ).toBe(1)
+    expect(section?.before).toContain('--remoteWrite.tmpDataPath=/tmpData')
+    expect(section?.after).toContain('--remoteWrite.tmpDataPath=/tmpData')
+  })
+
+  test('preserves rollback direction and operation counts', () => {
+    const result = helmPlanDiff(redisClusterRollbackPlan)
+
+    expect(result.description).toBe('Operation: Rollback')
+    expect(result.sections.map(({ operation }) => operation)).toEqual([
+      'update',
+      'delete',
+      'update',
+    ])
+    expect(result.summary).toMatchObject({
+      create: 0,
+      update: 2,
+      delete: 1,
+    })
+    expect(result.sections[0]?.before).toContain('image: redis:7.2.4-alpine')
+    expect(result.sections[0]?.after).toContain('image: redis:7.2.3-alpine')
+    expect(result.sections[1]?.before).toContain('kind: Service')
+    expect(result.sections[1]?.after).toBe('')
+    expect(result.sections[2]?.before).toContain('sentinel monitor mymaster')
+    expect(result.sections[2]?.after).not.toContain('sentinel')
   })
 
   test('preserves modified and error entries from v2 payloads', () => {
@@ -72,14 +110,10 @@ describe('helmPlanDiff', () => {
   test('strips ANSI sequences and searches all resource metadata', () => {
     const [section] = helmPlanDiff(vmagentSingleRemovalPlan).sections
 
-    expect(section?.title).toBe('metrics-agent')
+    expect(section?.title).toBe('vmagent')
+    expect(section?.description).toBe('Deployment · apps · observability')
     expect(section?.searchable).toEqual(
-      expect.arrayContaining([
-        'observability',
-        'metrics-agent',
-        'Deployment',
-        'apps/v1',
-      ])
+      expect.arrayContaining(['observability', 'vmagent', 'Deployment', 'apps'])
     )
   })
